requirements.txt
streamlit==1.30.0
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0
mypy
datasets # For Hugging Face datasets
Pillow # For image manipulation
pytest # For running tests
pytest-mock # For mocking in pytest
# Add versions for any OS-specific dependencies if needed, e.g., for faiss later

# Linters and Formatters
black==23.10.0
flake8==6.1.0

# Embeddings and Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
# Dependencies for Hugging Face datasets streaming
fsspec
aiohttp
